+++ go.mod
@@ -3,29 +3,22 @@ module github.com/mrodriguez3313/IPNSGoServer
 go 1.14
 
 require (
-	github.com/fsnotify/fsnotify v1.5.1 // indirect
+	github.com/btcsuite/btcd v0.22.0-beta // indirect
 	github.com/gammazero/deque v0.1.1
-	github.com/gopherjs/gopherjs v0.0.0-20190812055157-5d271430af9f // indirect
+	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/gorilla/mux v1.7.3
-	github.com/ipfs/go-datastore v0.5.1 // indirect
+	github.com/ipfs/go-cid v0.1.0 // indirect
 	github.com/ipfs/go-ipfs-api v0.3.0
-	github.com/ipfs/go-path v0.2.1 // indirect
-	github.com/ipfs/interface-go-ipfs-core v0.5.2
-	github.com/ipld/go-ipld-prime v0.14.2 // indirect
-	github.com/libp2p/go-libp2p v0.16.0 // indirect
+	github.com/kr/pretty v0.3.0 // indirect
 	github.com/libp2p/go-libp2p-core v0.12.0 // indirect
-	github.com/libp2p/go-libp2p-record v0.1.3 // indirect
 	github.com/multiformats/go-base32 v0.0.4 // indirect
 	github.com/multiformats/go-multiaddr v0.4.1 // indirect
-	github.com/robfig/cron v1.2.0 // indirect
-	github.com/smartystreets/assertions v1.0.1 // indirect
-	github.com/warpfork/go-testmark v0.9.0 // indirect
-	github.com/whyrusleeping/cbor-gen v0.0.0-20210219115102-f37d292932f2 // indirect
+	github.com/multiformats/go-multihash v0.1.0 // indirect
+	github.com/robfig/cron v1.2.0
 	github.com/whyrusleeping/tar-utils v0.0.0-20201201191210-20a61371de5b // indirect
-	go.uber.org/goleak v1.1.11 // indirect
-	go.uber.org/zap v1.19.1 // indirect
 	golang.org/x/crypto v0.0.0-20211202192323-5770296d904e // indirect
-	golang.org/x/lint v0.0.0-20201208152925-83fdc39ff7b5 // indirect
 	golang.org/x/sys v0.0.0-20211205182925-97ca703d548d // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 	lukechampine.com/blake3 v1.1.7 // indirect
 )
