--- misc/References/README.md
@@ -1,9 +1,0 @@
-# Note
-
-This folder holds old code to show different ways of how operations were done. Functions either won't work at all and just serve as examples. Or work on their own, but did not work with moving the API foward. There are comments indicating the operations of each. It is kept here to reference later on and show examples for operations like:
-
-- Generate ed25519 key from /go-libp2p-core/crypto library
-- Create a new keystore
-- Creating entry with IPNS library
-- Getting and saving all files/sub-files in a submitted directory.
-- In general, code used at some point.
