+++ src/ipfs.go
@@ -1,9 +1,11 @@
 package main
 
 import (
-	"fmt"
-	"os"
 	"context"
+	"fmt"
+	"net/http"
+	"os"
+	"strings"
 
 	shell "github.com/ipfs/go-ipfs-api"
 )
@@ -17,6 +19,87 @@ type ResolvedPath struct {
 	Path string
 }
 
+// Grab uploaded file and add to ipfs
+// returns ipfsPath (ipfsURI + CID)
+func AddFile(w http.ResponseWriter, r *http.Request) {
+	var dir = abs + "/Uploads"
+
+	FileName, err := saveFile(r, dir, MAX_UPLOAD_SIZE) // grab uploaded content & save to disk
+	if err != nil {
+		writeJSONError(w, "Error in saveContent", err)
+		return
+	}
+
+	fmt.Println("Adding to IPFS...")
+	ipfsPath, err := addToIPFS(dir+"/"+FileName, "") // add content file to IPFS
+	if err != nil {
+		writeJSONError(w, "Error in addToIPFS", err)
+		return
+	}
+	fmt.Println("ipfs Path: ", ipfsPath)
+	writeJSONSuccess(w, "Success - addFile", ipfsPath)
+}
+
+// Grab uploaded file and add to ipfs
+// returns ipfsPath (ipfsURI + CID)
+func addFolder(w http.ResponseWriter, r *http.Request) {
+	var dir = abs + "/Uploads"
+
+	// Check uploaded files/Dir is not bigger than MAX_UPLOAD_SIZE
+	r.Body = http.MaxBytesReader(w, r.Body, MAX_UPLOAD_SIZE)
+	if err := r.ParseMultipartForm(MAX_UPLOAD_SIZE); err != nil {
+		writeJSONError(w, "File too big. Max size is 3MB", err)
+		return
+	}
+
+	// The argument to FormFile must match the name attribute
+	// of the file input on the frontend
+	file, fileHeader, err := r.FormFile("files")
+	if err != nil {
+		writeJSONError(w, "Error accessing FormFile", err)
+		return
+	}
+	defer file.Close()
+
+	if msg, err := checkFileType(file, fileHeader); err != nil {
+		writeJSONError(w, msg, err)
+		return
+	}
+
+	dirPath, err := saveDir(file, fileHeader, dir, MAX_UPLOAD_SIZE) // grab uploaded content & save to disk
+	if err != nil {
+		writeJSONError(w, "Error in saveDir", err)
+		return
+	}
+
+	// fmt.Println("Dir created: ", dirPath)
+	cleanedFileName := cleanFileName(fileHeader.Filename)
+	fileName := removeExtenstion(cleanedFileName)
+	err = unzipSource(dirPath+"/"+cleanedFileName, dirPath+"/unzip/")
+	if err != nil {
+		writeJSONError(w, "Error in unzipSource", err)
+		return
+	}
+
+	// fmt.Println("fileName: ", fileName)
+	fmt.Println("Adding to IPFS...")
+	ipfsPath, err := addToIPFS(strings.Join([]string{dirPath, "/unzip/", fileName}, ""), "r") // add content file to IPFS
+	if err != nil {
+		writeJSONError(w, "Error in addToIPFS", err)
+		return
+	}
+
+	// Remove all the directories and files
+	// Using RemoveAll() function
+	err = os.RemoveAll(dirPath)
+	if err != nil {
+		writeJSONError(w, "Error in deleting directory", err)
+		return
+	}
+
+	fmt.Println("ipfs Path: ", ipfsPath)
+	writeJSONSuccess(w, "Success - addFolder", ipfsPath)
+}
 
 // Adds path to IPFS, given as a string
 // returns CID
@@ -31,7 +114,7 @@ func addToIPFS(path string, option string) (string, error) {
 			return "", err
 		}
 		ipfsPath = ipfsURI + cid
-		
+
 	} else {
 		fileReader, err := os.Open(path)
 		if err != nil {
@@ -52,7 +135,6 @@ func addToIPFS(path string, option string) (string, error) {
 	return ipfsPath, nil
 }
 
-
 // Custom publishing function returns the response object and error
 // Kept as close as possible to Publish method found at gh.com/go-ipfs-api
 func PublishToIPFS(contentHash string, key string) (*PublishResponse, error) {
@@ -70,8 +152,8 @@ func PublishToIPFS(contentHash string, key string) (*PublishResponse, error) {
 
 // This function is needed to let the world know your Record exists.
 func publishToIPNS(ipfsPath string, KeyName string) (*PublishResponse, error) {
-    pubResp, err := PublishToIPFS(ipfsPath, KeyName)
-    if err != nil {
+	pubResp, err := PublishToIPFS(ipfsPath, KeyName)
+	if err != nil {
 		fmt.Fprintf(os.Stderr, "Error in Publish: %s ", err)
 		return nil, err
 	}
@@ -82,11 +164,11 @@ func publishToIPNS(ipfsPath string, KeyName string) (*PublishResponse, error) {
 	}
 
 	fmt.Printf("\nresponse Name: %s\nresponse Value: %s\n", pubResp.Name, pubResp.Value)
-	
+
 	return pubResp, nil
 }
 
-// This function will resolve/download the content pointed to by the record. 
+// This function will resolve/download the content pointed to by the record.
 func resolve(ipnsKey string) (string, error) {
 	var path ResolvedPath
 	sh := shell.NewShell(localhost)
